requirements.txt
streamlit>=1.33.0
langchain>=0.2.11
langchain-core>=0.2.24
langchain-openai>=0.1.17
openai>=1.40.0
pydantic>=2.7.0
python-dotenv>=1.0.1
requests>=2.32.0
typer>=0.12.3
rich>=13.7.1
